% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gage_temperature_wqx}
\alias{gage_temperature_wqx}
\title{Temperature Gages for WQX data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 8 columns.
}
\usage{
gage_temperature_wqx
}
\description{
Temperature Gages for WQX data
}
\examples{
gage_temperature_wqx |> pillar::glimpse()
}
\keyword{datasets}
