% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_gage}
\alias{ph_gage}
\title{pH gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
A tibble with 124 rows and 8 columns
\itemize{
\item \code{location}: location associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{agency}: agency responsible for operating or maintaining the gage (e.g., USGS, State of Oregon Dept. of Environmental Quality,
Hoopa Valley Tribe (Tribal), Bureau of Reclamation, Klamath Tribes (Tribal), Yurok Tribe of the Yurok Reservation, California (Tribal),
California State Water Resources Control Board, National Park Service Water Resources Division, EPA National Aquatic Resources Survey (NARS))
\item \code{latitude}: latitude coordinate of the gage location
\item \code{longitude}: longitude coordinate of the gage location
\item \code{river_mile}: closest river mile location relative to river mouth or a reference point
\item \code{huc8}: 8-digit Hydrologic Unit Code identifying the sub-basin where the gage is located
}
}
\usage{
ph_gage
}
\description{
A dataset containing metadata for pH gages located across the Klamath Basin watershed.
Includes location, agency, and hydrologic unit code (HUC) information for each station.
}
\keyword{datasets}
