% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_gage}
\alias{temperature_gage}
\title{Temperature gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
A tibble with 94 rows and 8 columns
\itemize{
\item{\code{stream}: name of the stream or river where the gage is located
\item{\code{gage_name}: name of the gaging station
\item{\code{gage_id}: unique identifier for the gage
\item{\code{agency}: agency responsible for operating or maintaining the gage (e.g., USGS, State of Oregon Dept. of Environmental Quality,
Hoopa Valley Tribe (Tribal), USDA FS PIBO Monitoring Program)
\item{\code{latitude}: latitude coordinate of the gage location
\item{\code{longitude: longitude coordinate of the gage location
\item{\code{river_mile}: closest river mile location relative to river mouth or a reference point
\item{\code{huc8}: 8-digit Hydrologic Unit Code identifying the sub-basin where the gage is located
}
}
\usage{
temperature_gage
}
\description{
A dataset containing metadata for temperature gages located across the Klamath Basin watershed.
Includes location, agency, and hydrologic unit code (HUC) information for each station.
}
\keyword{datasets}
