% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_wqx}
\alias{temperature_wqx}
\title{All temperature data from WQX}
\format{
An object of class \code{data.frame} with 99076 rows and 8 columns.
}
\usage{
temperature_wqx
}
\description{
All temperature data from WQX
}
\examples{
temperature_wqx |> pillar::glimpse()
}
\keyword{datasets}
