% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgs_dam_removal_monitoring_layers}
\alias{usgs_dam_removal_monitoring_layers}
\title{Spatial layers from USGS dam removal monitoring in the Klamath Basin}
\format{
An object of class \code{list} of length 9.
}
\usage{
usgs_dam_removal_monitoring_layers
}
\description{
A named list of spatial (\code{sf}) layers pulled from the Klamath Dam Removal Monitoring Web Map hosted on ArcGIS Online.
Each element represents a layer of interest, such as reservoir footprints, sediment monitoring points, or dam sites.
}
\examples{
names(usgs_dam_removal_monitoring_layers)
usgs_dam_removal_monitoring_layers$geomorphic_reaches |> sf::st_geometry() |> plot()
}
\keyword{datasets}
