% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{do_gage}
\alias{do_gage}
\title{Dissolved oxygen gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 8 columns.
}
\usage{
do_gage
}
\description{
Dissolved oxygen gages in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
do_gage |> pillar::glimpse()
}
\keyword{datasets}
