% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_data}
\alias{flow_data}
\title{Publicly available flow data in the Klamath Basin watershed (HUC8 18010204)}
\format{
A tibble with 227,149 rows and 8 columns
\itemize{
\item \code{location}: location associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{variable_name}: variable measured: "flow"
\item \code{value}: observed value of the temperature (numeric)
\item \code{unit}: units of measurement: cfs
\item \code{statistic}: summary statistic type: mean"
\item \code{date}: date of the observation
}
}
\usage{
flow_data
}
\description{
Mean flow data from gages across the Klamath Basin. Raw data was obtained from USGS, WQX, USBR, and OWRD data portals.
}
\keyword{datasets}
