% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_data}
\alias{flow_data}
\title{Publicly available flow data in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{data.frame} with 114654 rows and 8 columns.
}
\usage{
flow_data
}
\description{
Publicly available flow data in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
flow_data |> pillar::glimpse()
}
\keyword{datasets}
