% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_gage}
\alias{flow_gage}
\title{Flow gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
A tibble with 117 rows and 8 columns
\itemize{
\item \code{location}: location associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{agency}: agency responsible for operating or maintaining the gage (e.g., USGS, Quartz Valley Indian Community of the Quartz Valley Reservation of California (Tribal),
Klamath Tribes (Tribal))
\item \code{latitude}: latitude coordinate of the gage location
\item \code{longitude}: longitude coordinate of the gage location
\item \code{river_mile}: closest river mile location relative to river mouth or a reference point
\item \code{huc8}: 8-digit Hydrologic Unit Code identifying the sub-basin where the gage is located
}
}
\usage{
flow_gage
}
\description{
A dataset containing metadata for flow gages located across the Klamath Basin watershed.
Includes location, agency, and hydrologic unit code (HUC) information for each station.
}
\keyword{datasets}
