% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_gage}
\alias{flow_gage}
\title{Flow gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 8 columns.
}
\usage{
flow_gage
}
\description{
Flow gages in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
flow_gage |> pillar::glimpse()
}
\keyword{datasets}
