% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_data}
\alias{ph_data}
\title{Publicly available pH data in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{data.frame} with 111594 rows and 8 columns.
}
\usage{
ph_data
}
\description{
Publicly available pH data in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
ph_data |> pillar::glimpse()
}
\keyword{datasets}
