% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_gage}
\alias{ph_gage}
\title{pH gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 8 columns.
}
\usage{
ph_gage
}
\description{
pH gages in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
ph_gage |> pillar::glimpse()
}
\keyword{datasets}
