% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_data}
\alias{temperature_data}
\title{Publicly available temperature data in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{data.frame} with 348642 rows and 8 columns.
}
\usage{
temperature_data
}
\description{
Publicly available temperature data in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
temperature_data |> pillar::glimpse()
}
\keyword{datasets}
