% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_gage}
\alias{temperature_gage}
\title{Temperature gages in the Klamath Basin watershed (HUC8 18010204)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 94 rows and 8 columns.
}
\usage{
temperature_gage
}
\description{
Temperature gages in the Klamath Basin watershed (HUC8 18010204)
}
\examples{
temperature_gage |> pillar::glimpse()
}
\keyword{datasets}
