% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_level_data}
\alias{water_level_data}
\title{Lake water surface elevation data from four different USGS gages across the Klamath Basin}
\format{
A tibble with 4 rows and 8 columns
\itemize{
\item \code{location}: location associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{variable_name}: variable measured: "water level"
\item \code{value}: observed value of the temperature (numeric)
\item \code{unit}: units of measurement: feet
\item \code{statistic}: summary statistic type: "mean"
\item \code{date}: date of the observation
}
}
\usage{
water_level_data
}
\description{

}
\keyword{datasets}
