% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_level_data}
\alias{water_level_data}
\title{Lake water surface elevation data from four different USGS gages, and one USBR monitoring site across the Klamath Basin}
\format{
A tibble with 30,084 rows and 8 columns:
\itemize{
\item \code{location}: location associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{variable_name}: variable measured: "water level"
\item \code{value}: observed value of the temperature (numeric)
\item \code{unit}: units of measurement: feet
\item \code{statistic}: summary statistic type: "mean"
\item \code{date}: date of the observation
}
}
\usage{
water_level_data
}
\description{
Lake water surface elevation (ft) data from the following gages:
\itemize{
\item 11505900
\item 11504300
\item 11505800
\item 11507001
\item usbr tule lake sump 1a
}
}
\keyword{datasets}
