% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_level_gage}
\alias{water_level_gage}
\title{A dataset containing metadata for water level gages located across the Klamath Basin watershed.
Includes location, agency, and hydrologic unit code (HUC8) information for each station.
\itemize{
\item 11505900
\item 11504300
\item 11505800
\item 11507001}
\format{
A tibble with 4 rows and 8 columns
\itemize{
\item \code{stream}: stream associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{variable_name}: variable measured: "water level"
\item \code{value}: observed value of the temperature (numeric)
\item \code{unit}: units of measurement: feet
\item \code{statistic}: summary statistic type: "mean"
\item \code{date}: date of the observation
}
}
\usage{
water_level_gage
}
\description{
A dataset containing metadata for water level gages located across the Klamath Basin watershed.
Includes location, agency, and hydrologic unit code (HUC8) information for each station.
\itemize{
\item 11505900
\item 11504300
\item 11505800
\item 11507001
}
\keyword{datasets}
