% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_level_gage}
\alias{water_level_gage}
\title{Metadata for water level gages in the Klamath Basin.}
\format{
A tibble with 5 rows and 7 columns:
\itemize{
\item \code{location}: location associated where data was collected/where gage is located
\item \code{gage_name}: name of the gaging station
\item \code{gage_id}: unique identifier for the gage
\item \code{agency}: agency responsible for operating or maintaining the gage
\item \code{latitude}: latitude coordinate of the gage location
\item \code{longitude}: longitude coordinate of the gage location
\item \code{huc8}: 8-digit Hydrologic Unit Code identifying the sub-basin where the gage is located
}
}
\usage{
water_level_gage
}
\description{
A dataset containing metadata for water level gages located
across the Klamath Basin watershed, including location, agency, and
Hydrologic Unit Code (HUC8) information for each station.
\itemize{
\item 11505900
\item 11504300
\item 11505800
\item 11507001
\item usbr tule lake sump 1a
}
}
\keyword{datasets}
